fix(dashpay): fix username UI crashes and support recovery during username creation  (#1326)
* fix: eliminate cast exception in BlockInfoActivity

* fix: remove unused fragments in Username Navigation

* fix: prevent NPE's

* chore: remove unnecessary code in UsernameRegistrationFragment

* chore: use dpp v1.5.1-SNAPSHOT

* fix: allow 128 char links for verification

* fix: simplify coinjoin system notification status updates

* fix: cancel work if databased are cleared

* fix: more screen fixes

* fix: allow credits to be used on WelcomeToDashPayFragment

* fix: use a double for mixing status

* fix: handle failed restore after failed request

* fix: CreateIdentityService fixes

+++ wallet/src/de/schildbach/wallet/service/BlockchainServiceImpl.java
@@ -224,7 +224,7 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
     private final Executor executor = Executors.newSingleThreadExecutor();
     private int syncPercentage = 0; // 0 to 100%
     private MixingStatus mixingStatus = MixingStatus.NOT_STARTED;
-    private Double mixingProgress = 0.0;
+    private double mixingProgress = 0.0;
     private ForegroundService foregroundService = ForegroundService.NONE;
 
     // Risk Analyser for Transactions that is PeerGroup Aware
@@ -1088,7 +1088,7 @@ public class BlockchainServiceImpl extends LifecycleService implements Blockchai
         final String message = getString(
                 R.string.coinjoin_progress,
                 getString(statusStringId),
-                mixingProgress.intValue(),
+                mixingProgress,
                 decimalFormat.format(MonetaryExtKt.toBigDecimal(mixedBalance)),
                 decimalFormat.format(MonetaryExtKt.toBigDecimal(totalBalance))
         );
